Changes to the leaflet control container
Changes the leaflet control container to add the tray position. This
moves the original UI elements from the top left to the bottom.

The tray allows for the user to hide the UI when they want to fully see
the map.

Also added a bunch of documentation comments to the code.